fix(e2e): scope the chromatic fixture to visual snapshot specs only
