+++ e2e/playwright.auth.config.ts
@@ -1,4 +1,3 @@
-import {type ChromaticConfig} from '@chromatic-com/playwright'
 import {defineConfig, devices} from '@playwright/test'
 
 const CI = process.env.CI === 'true'
@@ -20,7 +19,7 @@ const PORT = 3340
  *   pnpm build --filter=auth-test-studio...
  *   pnpm --filter auth-test-studio start --port 3340
  */
-export default defineConfig<ChromaticConfig>({
+export default defineConfig({
   testDir: './tests/auth',
   timeout: 60_000,
   fullyParallel: true,
@@ -31,9 +30,6 @@ export default defineConfig<ChromaticConfig>({
   retries: 1,
   reporter: [['list']],
   use: {
-    // Auth specs share the Chromatic-wrapped `test` from ./studio-test but
-    // are never uploaded to Chromatic — keep snapshotting off entirely.
-    disableAutoSnapshot: true,
     actionTimeout: 10_000,
     trace: 'on-first-retry',
     viewport: {width: 1728, height: 1000},
